Title: 앞에서부터 삭제하기 2, Time: 914ms, Memory: 35MB, Status: Passed - Codetree

+++ 250516/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -31,7 +31,7 @@ public class Main {
                 }
                 pq.poll();
             }
-                max = Math.max(max, (total - pq.peek()) / (count-1));
+                max = Math.max(max, (double)(total - pq.peek()) / (count-1));
             
         }
